% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaggregate.R
\name{aaggr}
\alias{aaggr}
\title{Aggregate the \code{.rda} files in a given directory}
\usage{
aaggr(directory)
}
\arguments{
\item{directory}{path to the directory to search.}
}
\value{
a matrix containing the contents of the \code{.rda} files
 in \code{directory}.
}
\description{
Read all the files in a directory ending in ".rda" and combine their contents
by either \code{cbind} or \code{rbind}. Assumes each \code{.rda} file contains
a single object. Assumes array inputs, produces array outputs. Combines by row.
}

