% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.glmnet.R
\name{AIC.glmnet}
\alias{AIC.glmnet}
\title{AIC from glmnet object}
\usage{
\method{AIC}{glmnet}(glm_fit)
}
\arguments{
\item{glm_fit}{object of type glmnet}
}
\description{
Calculates AIC modified by effective degrees of freedom from glmnet lasso
models
}
