% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{bootsample}
\alias{bootsample}
\title{Draw bootstrap samples}
\usage{
bootsample(n, b = 1000)
}
\arguments{
\item{n}{The size of the data set to draw samples from}

\item{b}{Number of bootstrap resamples to draw}
}
\value{
Returns an object of class "bootsample" with the following components
\item{samples}{a \code{b} by \code{n} matrix containing bootstrap samples}
\item{oob}{a list of b vectors that provide the samples that arent in the
 respective bootstrap samples}
}
\description{
Draws n samples in the range \code{1:n} with replacement. Provides the
out-of-bag samples for testing.
}
\examples{
data("iris")
nsamples <- nrow(iris)
bs <- bootsample(nsamples, 1)  # draw a single bootstrap sample

## split into training and test sets based on the bootstrap resampling
train <- iris[bs$sample[1, ], ]
test <- iris[bs$oob[[1]], ]
}

