% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.R
\name{calibration_curve}
\alias{calibration_curve}
\title{Calibration curve plot}
\usage{
calibration_curve(predictions, responses)
}
\arguments{
\item{predictions}{vector of predicted probabilities}

\item{responses}{vector of zeroes and ones denoting the truth}
}
\description{
Shows the calibration curve of a set of predictions. This is the ratio of the
smoothed histogram of predicted probabilities for the "ones" to the smoothed
histogram of all predicted probabilities. This estimates the proportion of class
 "one" that receives a given predicted probability. Ideally these should be
 equal.
}
