% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterkernel.R
\name{clusterkernel}
\alias{clusterkernel}
\title{Bagged cluster kernel}
\usage{
clusterkernel(x, ncluster, ninit = 100)
}
\arguments{
\item{x}{A matrix of data points, samples in the rows.}

\item{ncluster}{Maximum number of clusters to use (maximum \code{k} for
kmeans).}

\item{ninit}{Number of times to repeat the clusterings. As k-means has some
randomness to it, it might be a good idea to run it several times.}
}
\value{
A \code{nrow(x)} by \code{nrow(x)} matrix of pairwise similarities in
the range [0, 1]
}
\description{
Implements the bagged cluster kernel desctibed by Weston et al. Calculates a
distance metric between samples based on co-occurence in k-means clustering.
The algorithm clusters the data using a range of different values for
\code{k} in k-means and then uses the fraction of times two data points occur
in  the same cluster as their similarity.
}
\references{
Weston, J., Leslie, C., Ie, E., Zhou, D., Elisseeff, A., & Noble, W. S. (2005).
 Semi-supervised protein classification using cluster kernels.
 Bioinformatics, 21(15), 3241-3247.
}
