% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/einr.R
\docType{package}
\name{einr}
\alias{einr}
\alias{einr-package}
\title{einr: a collection of functions &c.}
\description{
This package collects functions that I have implemented either because I
keep needing the same thing over and over or because I was curious about
some method that I couldn't find implemented elsewhere. Or because I for some
reason wanted to implement it myself even though it was already out there.
}

