% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmvector.R
\name{glmvector}
\alias{glmvector}
\title{GLM vector}
\usage{
glmvector(x, y, family = gaussian, swapxy = F)
}
\arguments{
\item{x}{A predictor matrix with samples in the rows.}

\item{y}{A response vector.}

\item{family}{A description of the error distribution/link function for the
glm. For more info see the documentation for \code{glm}.}

\item{swapxy}{A logical value indicating that instead of regressing \code{y}
 on each column in \code{x}, you would like to regress the columns in \code{x}
 on \code{y}.}
}
\value{
an object of class "glmvector" with the following components:
\item{glms}{a list of univariate glms corresponding to the predictors in
   \code{x}.}
\item{pvalues}{the p-values from the univariate regressions.}
\item{slopes}{the estimated slopes from the regressions.}
\item{names}{predictor names. Basically \code{colnames(x)}}
}
\description{
Fits univariate GLMs for each of the predictors in your data set. Assumes
that each column is a predictor.
}
\seealso{
\code{\link{glm}}
}

