% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestvector.R
\name{hist.ttestvector}
\alias{hist.ttestvector}
\title{Historgram of unadjusted p-values for \code{ttestvector}}
\usage{
\method{hist}{ttestvector}(x, ...)
}
\arguments{
\item{x}{n object of class \code{ttestvector}}

\item{...}{extra arguments passed on to \code{hist}}
}
\description{
Historgram of unadjusted p-values for \code{ttestvector}
}
\seealso{
\code{\link{ttestvector}}, \code{\link{hist}}
}

