% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leekasso.R
\name{leekasso}
\alias{leekasso}
\title{Leekasso feature selection}
\usage{
leekasso(x, y)
}
\arguments{
\item{x}{A predictor matrix.}

\item{y}{A response vector.}
}
\value{
Vector containing the indexes of the top ten predictors.
}
\description{
Jeff Leek's leekasso for p-value-based feature selection. Performs univariate
regressions of the response on each of the predictors and selects the ten
predictors with the lowest p-values.
}

