% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmvector.R
\name{plot.glmvector}
\alias{plot.glmvector}
\title{Plot for glmvector objects}
\usage{
\method{plot}{glmvector}(x, ...)
}
\arguments{
\item{x}{a \code{glmvector} object.}

\item{...}{other parameters passed on to \code{plot()}}
}
\description{
Creates a "volcano plot" for the glm vector: plots slope as a function of
negative log p-value for each \code{glm} in the vector
}

