% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.density.R
\name{predict.density}
\alias{predict.density}
\title{Predict method for density estimates}
\usage{
\method{predict}{density}(obj, oldx, newx)
}
\arguments{
\item{obj}{an object of class "density"}

\item{oldx}{the data from which the density was estimated. this is needed
because the \code{density} method does not keep the original data}

\item{newx}{the values at which you want to compute the density}
}
\value{
a vector of density values at \code{newx}
}
\description{
Predicted values based on density object
}
\seealso{
\code{\link{density}}
}
