% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaggregate.R
\name{rdaggregate}
\alias{rdaggregate}
\title{Aggregate the \code{.rda} files in a given directory}
\usage{
rdaggregate(directory, byrow = T)
}
\arguments{
\item{directory}{path to the directory to search.}

\item{byrow}{combine by row or by column. Defaults to \code{TRUE}.}
}
\value{
a matrix or data frame containing the contents of the \code{.rda} files
 in \code{directory}.
}
\description{
Read all the files in a directory ending in ".rda" and combine their contents
by either \code{cbind} or \code{rbind}. Assumes each \code{.rda} file contains
a single object.
}

