% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_einr.R
\name{theme_einr}
\alias{theme_einr}
\title{Theme for ggplot}
\usage{
theme_einr(base_size = par()$ps, base_family = par()$family)
}
\description{
This is at the moment a small fix to ggthemes' theme_par(). The only change is
that legends take their font size from the cex parameter.
}
