% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestvector.R
\name{ttestvector}
\alias{ttestvector}
\title{T-test vector}
\usage{
ttestvector(x, y = NULL, adjust = TRUE, ...)
}
\arguments{
\item{x}{First group. Samples in rows}

\item{y}{Second group. Samples in rows. If \code{y=NULL} (default), a
one-sample test is performed exactly as in \code{t.test}}

\item{adjust}{Adjust for multiple testing? Boolean indicating whether the
FDR should be controlled by Benjamini-Hochberg correction using \code{p.adjust}.}

\item{...}{Additional arguments passed to \code{t.test}}
}
\value{
an object of class "ttestvector" that contains the following elements:
\item{tests}{a list that contains the \code{htest} objects corresponding to
 each call to \code{t.test}}
\item{pvalues}{an array of the adjusted p-values corresponding to each test. If
 \code{adjust} was \code{FALSE}, these will be the same as \code{rawpvalues}}
\item{rawpvalues}{an array of the unadjusted p-values corresponding to each test.}
\item{adjusted}{logical indicating whether p-values were adjusted}
}
\description{
Runs \code{t.test} for each column in the two input sets
}
\seealso{
\code{\link{t.test}}, \code{\link{p.adjust}}
}

