% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bootsample}
\alias{bootsample}
\title{Draw bootstrap samples}
\usage{
bootsample(n, b = 1000)
}
\arguments{
\item{n}{The size of the data set to draw samples from}

\item{b}{Number of bootstrap resamples to draw}
}
\value{
Returns an object of class "bootsample" which is a list of \code{b} lists
with the following two attributes:
\item{sample}{a \code{n}-vector containting a bootstrap sample.}
\item{oob}{a vector that contains the data points not in \code{sample}. Can
for eg be used as test set.}
}
\description{
Draws n samples in the range \code{1:n} with replacement. Provides the
out-of-bag samples for testing.
}
\examples{
data("iris")
nsamples <- nrow(iris)
bs <- bootsample(nsamples, 1)  # draw a single bootstrap sample

## split into training and test sets based on the bootstrap resampling
train_idx <- bs[[1]]$sample
test_idx <- bs[[1]]$oob
train <- iris[train_idx, ]
test <- iris[test_idx, ]
}
