% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hyperboxplot.R
\name{hyperboxplot}
\alias{hyperboxplot}
\title{Hyperboxplot: for when you have too many dimensions for regular boxplots}
\usage{
hyperboxplot(x)
}
\arguments{
\item{x}{A matrix from which the plot is produced. Each column is a separate
group or dimension.}
}
\value{
None.
}
\description{
This is a very naive reimplementation of Karl Broman's \code{manyboxplot},
which see \url{https://kbroman.wordpress.com/2012/04/25/microarrays-suck/}.
}

