% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leekasso.R
\name{leekasso}
\alias{leekasso}
\title{Leekasso regression}
\usage{
leekasso(x, y, family = gaussian)
}
\arguments{
\item{x}{A predictor matrix with samples in the rows.}

\item{y}{A response vector.}

\item{family}{A description of the error distribution/link function for the
glm. For more info see the documentation for \code{glm}.}
}
\value{
\code{leekasso} returns an object of class "leekasso" with the
following components:
\item{model}{a glm object fit with the best ten predictors.}
\item{pvals}{the p-values from the univariate regressions.}
\item{predictors}{the indexes of the best predictors.}
\item{x}{the data frame used to fit the model.}
\item{y}{the responses used to fit the model.}
}
\description{
Jeff Leek's leekasso for p-value-based feature selection. Performs univariate
regressions of the response on each of the predictors and selects the ten
predictors with the lowest p-values. Then regressses the response on these
ten predictors.
}
\seealso{
\code{\link{glm}}
}
