% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ttestvector.R
\name{ttestvector}
\alias{ttestvector}
\title{T-test vector}
\usage{
ttestvector(x, y, adjust = TRUE, ...)
}
\arguments{
\item{x}{First group. Samples in rows}

\item{y}{Second group. Samples in rows}

\item{adjust}{Adjust for multiple testing? Boolean indicating whether the
FDR should be controlled by Benjamini-Hochberg correction using \code{p.adjust}.}

\item{...}{Additional arguments passed to \code{t.test}}
}
\value{
an object of class "ttestvector" that contains the following elements:
\item{tests}{a list that contains the \code{htest} objects corresponding to
 each call to \code{t.test}}
\item{pvalues}{an array of the p-values corresponding to each test. If
 \code{adjust} was \code{TRUE}, these will be adjusted to control FDR}
\item{adjusted}{logical indicating whether p-values were adjusted}
}
\description{
Runs \code{t.test} for each column in the two input sets
}
\seealso{
\code{\link{t.test}}, \code{\link{p.adjust}}
}

